Add a test for submitting blank name
The result you are presented with when you've submitted a blank name is "You've submitted:", so I had assumed it allowed submission of empty strings. What I inteded to do here was create a failing test and then fix it. However, it turns out that even though it presents that message, it doesn't actually save it, so you don't end up with a blank string in the database. Handy!

I'm not sure it's worth changing the message you see, tbc.