This causes the box to hang
For some reason, this line is now causing the box to hang on `vagrant up`. This may be related to an upgrade to MacOS Catalina although more likely is an upgrade to VirtualBox 6, though I did not upgrade Virtualbox between this working and not working, so it's a mystery to me what caused the problem to start to manifest.

What the line I'd added does is disconnect a serial port (uart1) because there is an issue with the ubuntu vagrant boxes such that they try to connect to that port even if it's not connected and then print the log file when it fails (not printing the log file is the end result I was aiming for).

Removing this line means the log file is again created but I can at least connect.

https://github.com/hashicorp/vagrant/issues/10578