Make test database support unicode
Interesting situation here:

On the server, a name with unicode characters in it, e.g. Karol Józef Wojtyła, causes an error.
When running the app locally, it works.
When adding direct to the local database, it works.
When running the unit tests, it fails.

Got to the bottom of it. The local database server is being created with a utf-8 character set (https://github.com/annashipman/remotehat/blob/master/Vagrantfile#L28) but even though the default storage engine in InnoDB which I understood to default to utf-8, the test database is not being created with a utf-8 character set. This commit fixes that.

Additionally, the database on the server will have to be changed, but as this is currently is done manually, there will be no corresponding commit.