Add table for Rounds
Not sure if this might be constants, but I need the round accessible.

I do not want a session because that is per site per visitor
Apparently there is no state shared between views as they probably run in different threads, so a global will not work.
So Stack Overflow suggests storing in the db (or hidden field in form, or cookie. I prefer db for this, I think).