Get the current round from the database
This is a bit weirdly written. Initially I had `current_round = Round.objects.get().current_round` and `current_round` everywhere else, but then when I wanted to update the current round in the database, I had to get a new object, as I only had the text. I'm not sure what best practice is here and/or whether I should just resolve this by renaming one of the `current_round`s... but this works for now.

It does depend on the database being set up intially with a current round of 1; I'll do that in a later commit.